+++ db/migrate/20240608160742_create_good_job_labels_index.rb
@@ -8,14 +8,12 @@ class CreateGoodJobLabelsIndex < ActiveRecord::Migration[7.1]
       dir.up do
         unless connection.index_name_exists?(:good_jobs, :index_good_jobs_on_labels)
           add_index :good_jobs, :labels, using: :gin, where: "(labels IS NOT NULL)",
-            name: :index_good_jobs_on_labels, algorithm: :concurrently
+                                         name: :index_good_jobs_on_labels, algorithm: :concurrently
         end
       end
 
       dir.down do
-        if connection.index_name_exists?(:good_jobs, :index_good_jobs_on_labels)
-          remove_index :good_jobs, name: :index_good_jobs_on_labels
-        end
+        remove_index :good_jobs, name: :index_good_jobs_on_labels if connection.index_name_exists?(:good_jobs, :index_good_jobs_on_labels)
       end
     end
   end
